+++ src/app/settings/dialogs/access-token-dialog/access-token-dialog.component.ts
@@ -1,0 +1,29 @@
+import { Component, Inject } from '@angular/core'
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog'
+import { faCode } from '@fortawesome/free-solid-svg-icons'
+import { HandleSubscription } from 'common/handle-subscription'
+
+@Component({
+  selector: 'app-site-code-dialog',
+  templateUrl: './access-token-dialog.component.html',
+  styleUrls: ['./access-token-dialog.component.scss'],
+})
+export class AccessTokenDialogComponent extends HandleSubscription {
+  faCode = faCode
+  loadingInfo: boolean = true
+
+  constructor (
+    public dialogRef: MatDialogRef<AccessTokenDialogComponent>,
+    @Inject(MAT_DIALOG_DATA) public data: string,
+  ) {
+    super()
+  }
+
+  copyCode (elementId: string): void {
+    const input = <HTMLInputElement>document.getElementById(elementId)
+    input.focus()
+    input.select()
+    document.execCommand('copy')
+    input.setSelectionRange(0, 0)
+  }
+}