+++ src/app/store/settings/settings.effects.ts
@@ -4,15 +4,24 @@ import { Observable, of as observableOf, timer as observableTimer } from 'rxjs'
 import { catchError, map, switchMap, takeUntil } from 'rxjs/operators'
 import { USER_LOG_OUT_SUCCESS } from 'store/auth/auth.actions'
 import {
+  ADD_ACCESS_TOKEN,
+  AddAccessToken,
+  AddAccessTokenSuccess,
   CANCEL_AWAITING_TRANSACTION,
   CancelAwaitingTransaction,
   CancelAwaitingTransactionSuccess,
+  DELETE_ACCESS_TOKEN,
   DELETE_REF_LINK,
+  DeleteAccessToken,
+  DeleteAccessTokenSuccess,
   DeleteRefLink,
   DeleteRefLinkSuccess,
+  GET_ACCESS_TOKENS,
   GET_BILLING_HISTORY,
   GET_CURRENT_BALANCE,
   GET_REF_LINKS,
+  GetAccessTokens,
+  GetAccessTokensSuccess,
   GetBillingHistory,
   GetBillingHistoryFailure,
   GetBillingHistorySuccess,
@@ -83,6 +92,42 @@ export class SettingsEffects {
         )
       )
     ))
+
+  getAccessTokens$ = createEffect(() => this.actions$
+    .pipe(
+      ofType<GetAccessTokens>(GET_ACCESS_TOKENS),
+      switchMap(() => this.common.getAccessTokens()
+        .pipe(
+          map(accessTokens => new GetAccessTokensSuccess(accessTokens)),
+          catchError(error => observableOf(new ShowDialogOnError(`Token fetch failed. Error code: ${error.status}`)))
+        )
+      )
+    )
+  )
+
+  addAccessToken$ = createEffect(() => this.actions$
+    .pipe(
+      ofType<AddAccessToken>(ADD_ACCESS_TOKEN),
+      switchMap(action => this.common.addAccessToken(action.payload)
+        .pipe(
+          map(accessToken => new AddAccessTokenSuccess(accessToken)),
+          catchError(error => observableOf(new ShowDialogOnError(`Token creation failed. Error code: ${error.status}`)))
+        )
+      )
+    )
+  )
+
+  deleteAccessToken$ = createEffect(() => this.actions$
+    .pipe(
+      ofType<DeleteAccessToken>(DELETE_ACCESS_TOKEN),
+      switchMap(action => this.common.deleteAccessToken(action.payload)
+        .pipe(
+          map(() => new DeleteAccessTokenSuccess(action.payload)),
+          catchError(error => observableOf(new ShowDialogOnError(`Token deletion failed. Error code: ${error.status}`)))
+        )
+      )
+    )
+  )
 
   getRefLinks$ = createEffect(() => this.actions$
     .pipe(